parameterize "notes" (big refactor)
there is no longer "freq" and "note_on" params hard coded into the API.
instead, each module has a custom struct of params it accepts from
notes. these can be floats, bools, anything.

(actually, i had to hardcode "note_on" in one place, but hope i can
remove that later.)

the drawback of this is that you often need to take an extra step to
convert notes from one param type to another, which is hard to do
elegantly in zig.

also: modules are no longer initialized with any values that should come
from the notes. modules are basically inactive until the first notes
arrives. this is nice.

also: triggers are once again baked into modules. i don't like this, but
i'll keep thinking about it. it's a tradeoff between boilerplate for the
caller, and coupling in the module structs.

todo: curves need attention, as usual.