improve delay so that you can add effects in the feedback
delay doesn't follow the "module" api anymore, it's a unique object
now. example_delay has a few custom modules created using it.
the caller has to do a fair bit of work, but this was necessary to allow
effects to be added. there is also some ugly looping required for the
caller in order to make the delay work when DELAY_SAMPLES is less than
the mix buffer length.

example_delay now includes a low-pass filter in the echo feedback.