remove DynamicNoteTracker from ImpulseQueue
this was not needed. Triggerable is already keeping track of the
currently playing note.

hopefully this paves the way for getting rid of the linked lists (the
only reason i used them was for the convenience of "pushing" the current
note to the head, which is no longer needed, at least in ImpulseQueue).