refactor triggers - cleaner and more flexible (but more boilerplate to use)
after trying a bunch of things i've settled on this basic pattern.

- change Triggerable to Trigger. no longer contains the module. exists beside it
- paintFromImpulses removed, replaced with iterator pattern. caller must call into the module's paint method
- "reset" methods removed, replaced with "note_id_changed" argument (which paint methods can check in order to reset themselves)
- added a struct called "Span" which is now required to pass into every painting-related function. this replaces the comptime slicing magic i had before

it now takes more lines of code to use the library, but i'm fine with this because it's now cleaner and more explicit, and i got rid of the static [32] arrays required to map note params. this iterator pattern is probably the best i can do in a language without closures etc. all in all it doesn't feel too bad.

what i'm least happy with is the requirement to create a block scope to contain iterator variables, if you don't want them entering the outer scope (and colliding with other iterators you may have), but this is a general zig thing.