consolidate "line painting" (interpolation) code
used by portamento and envelope modules

- remove old contents of paint_line.zig
- move `Painter` from mod_envelope.zig to paint_line.zig
- example_portamento now uses a simple, clear oscillator instead
  of a fuzzy noise effect. so you can hear the effect better

the "CatchUp" style of interpolation no longer exists. i don't
think we need it. it was useless for ADSR envelopes because it
never actually hits its goal, just approaches it. it was also used
by portamento, but portamento seems to work fine with the Painter
thing