example modules specify exactly what outputs mean
example modules must set "output_audio" to either mono or stereo, and
specify which outputs to use. they can also set "output_visualize" to
provide input to visualizers (it used to be hacked to just the first
output).

oscilloscope can now be synced to a continuous signal with
"output_sync_oscilloscope". i changed a few examples, adding an extra
output to be used as the sync signal. (for other examples, oscilloscope
won't sync).

also added a new example "vibrato" just to demonstrate this.

unfortunately, no oscilloscope syncing in example_script_runtime because
it uses polyphony. i might take polyphony out or make a copy of this
example without polyphony. (eventually it won't matter once polyphony
and multiple outputs is built in to the scripting language.)