zangscript: fix bad assertion
the PrintHelper.deinit method was asserting that indentation state has
been cleaned up. this is bad because deinit is called with defer, which
means it's called even in error situations, where indentation state will
obviously not have been finished.

so replace it with a method called 'finish' and call it normally,
without defer