refactor codegen: simplify module resolution
don't get the list of fields in parse, do it in codegen instead. now
modules are just another kind of value, and the `()` call syntax is part
of expectExpression in general.

in codegen, get rid of all the "visitor" stuff i had before. now that
modules are just values, the code i have for resolving define-anywhere
globals takes care of that.